2021: Improved runner to prompt for day
Improved the test runner to prompt for which day to run if it's not
given.

The primary motivation for this is it makes it easier to use the
debugger in vscode, as I don't have to mess with the launch.json each
time to set which day I'd like it to run.